Created header, recipe item and fab to create new recipes

+++ app/src/main/java/com/example/recipesapp/data/Ingredient.kt
@@ -2,6 +2,6 @@ package com.example.recipesapp.data
 
 data class Ingredient(
     var name: String = "",
-    var quantity: Long? = null,
+    var quantity: Double? = null,
     var unit: String? = null
 )
